feat(webpack): add hot module reloading
Closes DCOS-21844
