+++ .babelrc
@@ -1,5 +1,5 @@
 {
-  "plugins": ["transform-runtime"],
+  "plugins": ["transform-runtime", "react-hot-loader/babel"],
   "presets": ["es2015", "stage-3", "react"]
 }
