+++ package.json
@@ -135,7 +135,7 @@
     "purify-css": "1.2.5",
     "raml-validator-loader": "0.1.11",
     "react-addons-test-utils": "15.6.2",
-    "react-hot-loader": "1.3.1",
+    "react-hot-loader": "4.0.1",
     "react-test-renderer": "15.4.1",
     "rxjs-marbles": "2.3.1",
     "source-map-loader": "0.1.5",
@@ -144,13 +144,13 @@
     "stylelint": "7.2.0",
     "stylelint-config-dcos": "0.0.3",
     "svg-sprite": "1.3.1",
-    "typescript": "2.7.2",
     "ts-loader": "3.5.0",
     "tslint": "5.9.1",
     "tslint-config-prettier": "1.10.0",
     "tslint-eslint-rules": "5.1.0",
     "tslint-plugin-prettier": "1.3.0",
     "tslint-react": "3.5.1",
+    "typescript": "2.7.2",
     "uglifyjs-webpack-plugin": "1.2.0",
     "vinyl": "1.1.1",
     "webpack": "3.10.0",
